+++ BE/exceed/src/main/java/com/gaebaljip/exceed/common/annotation/AuthenticationMemberIdArgumentResolver.java
@@ -1,6 +1,6 @@
 package com.gaebaljip.exceed.common.annotation;
 
-import com.gaebaljip.exceed.security.CustomUsernamePasswordAuthenticationToken;
+import com.gaebaljip.exceed.security.domain.CustomUsernamePasswordAuthenticationToken;
 import org.springframework.core.MethodParameter;
 import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.security.core.Authentication;
